MAINT: deleting hyperbolic_interface_laws.py
